requirements.txt
python-dotenv
requests
pandas
psycopg2-binary